Improve the way the theme loads translations to not clash with pro
This is a quick fix to prevent the theme unintentionally removing the
pro translations from the translation repository chain. It preserves -
or more correctly, recreates - the original loading order from the app
such that theme translations override pro which overrides the core app.

Ideally we should rethink this so that the theme doesn't have to know
everything about what repositories are supposed to be available and to
avoid building and loading the repositories twice over.